move menu-dialog from src/shared/auth/dialogs to src/app/core/dialog/menu
