+++ src/app/core/components/header/header.component.ts
@@ -8,7 +8,7 @@ import { UserService } from '../../services/user/user.service';
 import { User } from '../../model/user';
 import { DeleteUserDialogComponent } from '../../dialogs/delete-user-dialog/delete-user-dialog.component';
 import { LoginComponent } from '../../../shared/auth/dialogs/login/login.component';
-import { MenuDialogComponent } from 'src/app/shared/auth/dialogs/menu/menu.component';
+import { MenuDialogComponent } from '../../dialogs/menu-dialog/menu.component';
 import { SignupComponent } from '../../../shared/auth/dialogs/signup/signup.component';
 import { ConfirmComponent } from '../../../shared/auth/dialogs/confirm/confirm.component';
 import { LoginResult } from '../../../shared/auth/model/login-result';
