+++ src/app/core/dialogs/menu-dialog/menu.component.scss
@@ -1,5 +1,5 @@
 @charset "utf-8";
-@import "../../../../../assets/scss/_initializer.scss";
+@import "../../../../assets/scss/_initializer.scss";
 
 .menu-dialog-header{
   display: flex;
